+++ src/containerapp/HISTORY.rst
@@ -2,6 +2,18 @@
 
 Release History
 ===============
+Upcoming
++++++++
+* 'az containerapp create/update': --yaml support properties for api-version 2022-10-01 (e.g. exposedPort,clientCertificateMode,corsPolicy)
+* 'az containerapp env update': fix bugs in update environment.
+* Fix YAML create with user-assigned identity
+* Fix polling logic for long running operations.
+* Add `az containerapp ingress update` Command to Update Container App Ingress
+
+0.3.24
+++++++
+* Decouple with the `network` module.
+
 0.3.23
 ++++++
 * BREAKING CHANGE: 'az containerapp env certificate list' returns [] if certificate not found, instead of raising an error.